Move Plutus specific testing tools to `cardano-ledger-core:testlib`
* Remove `Test.Cardano.Ledger.Alonzo.CostModel`
* Add `Cardano.Ledger.Plutus` that re-exports all plutus related
  functionality
* Add a new module for Plutus tests and include one property test
