+++ eras/alonzo/impl/CHANGELOG.md
@@ -53,6 +53,7 @@
 
 ### `testlib`
 
+* Remove `Test.Cardano.Ledger.Alonzo.CostModel` in favor of `Test.Cardano.Ledger.Plutus`
 * Add `mkPlutusScript'`
 * Add `alwaysSucceedsLang` and `alwaysFailsLang`
 * Change `alwaysSucceeds` and `alwaysFails` to accept the language version at the type level.
