+++ libs/cardano-ledger-core/cardano-ledger-core.cabal
@@ -51,6 +51,7 @@ library
         Cardano.Ledger.Orphans
         Cardano.Ledger.PoolDistr
         Cardano.Ledger.PoolParams
+        Cardano.Ledger.Plutus
         Cardano.Ledger.Plutus.CostModels
         Cardano.Ledger.Plutus.Data
         Cardano.Ledger.Plutus.ExUnits
@@ -199,6 +200,7 @@ test-suite tests
         Test.Cardano.Ledger.AddressSpec
         Test.Cardano.Ledger.BaseTypesSpec
         Test.Cardano.Ledger.BinarySpec
+        Test.Cardano.Ledger.PlutusSpec
         Test.Cardano.Ledger.UMapSpec
 
     default-language: Haskell2010
