+++ libs/cardano-ledger-core/src/Cardano/Ledger/Plutus.hs
@@ -1,0 +1,15 @@
+module Cardano.Ledger.Plutus (
+  module Cardano.Ledger.Plutus.CostModels,
+  module Cardano.Ledger.Plutus.Data,
+  module Cardano.Ledger.Plutus.ExUnits,
+  module Cardano.Ledger.Plutus.Language,
+  module Cardano.Ledger.Plutus.TxInfo,
+  module Cardano.Ledger.Plutus.Evaluate,
+) where
+
+import Cardano.Ledger.Plutus.CostModels
+import Cardano.Ledger.Plutus.Data
+import Cardano.Ledger.Plutus.Evaluate
+import Cardano.Ledger.Plutus.ExUnits
+import Cardano.Ledger.Plutus.Language
+import Cardano.Ledger.Plutus.TxInfo
