+++ libs/cardano-ledger-core/test/Main.hs
@@ -3,6 +3,7 @@ module Main where
 import qualified Test.Cardano.Ledger.BaseTypesSpec as BaseTypesSpec
 import qualified Test.Cardano.Ledger.BinarySpec as BinarySpec
 import Test.Cardano.Ledger.Common
+import qualified Test.Cardano.Ledger.PlutusSpec as PlutusSpec
 import qualified Test.Cardano.Ledger.UMapSpec as UMapSpec
 
 main :: IO ()
@@ -12,3 +13,4 @@ main =
       BaseTypesSpec.spec
       BinarySpec.spec
       UMapSpec.spec
+      PlutusSpec.spec
